vexus_hub/requirements.txt
# Core
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
Flask-WTF==1.1.1
python-dotenv==1.0.0

# Banco de Dados
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9 # Para PostgreSQL
# ou mysqlclient==2.2.0 # Para MySQL

# APIs e Integrações
requests==2.31.0
openai==0.28.1
twilio==8.9.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
facebook_business==18.0.1
pytz==2023.3

# Background Jobs
celery==5.3.4
redis==5.0.1

# Dashboard e Gráficos
pandas==2.1.3
numpy==1.24.3
plotly==5.18.0

# Utilitários
python-dateutil==2.8.2
phonenumbers==8.13.27
pyjwt==2.8.0
bcrypt==4.1.2
Werkzeug==2.3.7

# Segurança (Fase 2)
email-validator==2.1.0
pydantic==2.5.3
cryptography==42.0.4
bandit==1.7.6
safety==2.3.5